requirements.txt
torch~=2.8.0+cu126
opencv-python~=4.12.0.88
numpy~=2.1.3
fastapi~=0.116.1
torchvision~=0.23.0
pillow~=11.3.0
scikit-learn~=1.7.1
tqdm~=4.67.1
matplotlib~=3.10.0